+++ package.json
@@ -3,6 +3,9 @@
   "version": "2.0.0",
   "description": "fake patch on packageJson file",
   "main": "index.js",
+  "bin": {
+    "pseudo-patch": "./index.js"
+  },
   "repository": {
     "type": "git",
     "url": "https://github.com/tomdaniels/pseudo-patch.git"
@@ -24,5 +27,8 @@
     "formatting"
   ],
   "author": "Tom Daniels",
-  "license": "ISC"
+  "license": "ISC",
+  "dependencies": {
+    "commander": "^8.0.0"
+  }
 }
